Quantize relative priority
This adjusts the node request relative priority system to become
more coarse as queues become longer.

Currently in the case of a very long queue we might revise the
pending node requests of every item in the queue each time a change
is dequeued.  This might be useful for a change near the head of
the queue since it could cause it to receive nodes faster, but
we don't need to update the node request for jobs with the 100th
change in the queue on every update.

To avoid all of the ZK writes generated by these updates, we will
quantize the relative priority logarithmically.  So changes near
the head behave as before, but after 10 changes back we decrease
the resolution to buckets of 10, and after 100 changes, buckets of
100.

Change-Id: I63c4b5d08f0d2608bc54094ade7d46d0720ab25e