Update patch set 38
Patch Set 38:

Note that the display at the top is not only input, but also information display -- it tells users what the timezone is currently set to.  Other applications do this too (consider grafana -- https://grafana.opendev.org/d/5Imot6EMk/zuul-status?orgId=1 ).  And this means that we don't need to add the timezone to every field that displays a date/time.

So if there is a justification for this change, it should probably address that issue too.

Patch-set: 38