Add support for filesets in job configuration
Given a project repository with folders A and B and corresponding job:

- job:
    name: Job_A
    files:
     - A/.*
    irrelevant-files:
     - .*\.py$

Job_A generally runs for changes in folder A unless only .py files are
changed. Given the following set of modified files:

A/a.py
B/b.cpp

The expectation is that Job_A is not executed because no relevant
files are modified. But surprisingly Job_A is executed unnecessarily.
This bechavior is caused by the matching logic of the files and
irrelevant-files matcher and the fact that it is evaluated based on
all modified files instead of only the job's files. To better
understand let's look at what happens in the Job.changeMatchesFiles
method:
 - if self.file_matcher and not self.file_matcher.matches(change):
      return False # file_matcher matches irrelevant file A/a.py so
                   # the job is not skipped here.
 - if (self.irrelevant_file_matcher and
       self.irrelevant_file_matcher.matches(change)):
      return False # irrelevant_file_macher doesn't match B/b.cpp so
                   # the job is also not skipped here.
 - return True # As a result the job is run.

Also see the difference in behavior in following tests:
test_irrelevant_files_one_not_included_one_excluded_runs_job vs.
test_fileset_one_not_included_one_excluded_skips_job

This change proposes an alternative and possibly more intuitive way to
define relevant files for a job inspired by ant filesets:
https://ant.apache.org/manual-1.9.x/Types/fileset.html

For each job a fileset can be configured by defining included and
excluded files (each can be a list of regexes), for example:

- job:
    name: Job_A
    fileset:
      includes:
       - A/.*
      excludes:
       - .*\.py$

To decide whether a job must be executed for a particular change this
fileset filter is applied to each modified file of the change
separately. If a file is included and not excluded then it is added to
the job's fileset. The job is only executed if its final fileset is
not empty.

files / irrelevant-files behavior is kept unchanged for backwards-
compatibility.

Change-Id: I0a71f4afc82b6a736b1ab13fc7862f23167845f7