Add support for filesets in job configuration
This change proposes an alternative and possibly more intuitive way to
define relevant files for a job inspired by ant filesets:
https://ant.apache.org/manual-1.9.x/Types/fileset.html

For each job a fileset can be configured by defining included and
excluded files (each can be a list of regexes), for example:

- job:
    name: Job_A
    fileset:
      includes:
       - A/.*
      excludes:
       - .*\.py$

To decide whether a job must be executed for a particular change this
fileset filter is applied to each modified file of the change
separately. If a file is included and not excluded then it is added to
the job's fileset. The job is only executed if its final fileset is
not empty.

files / irrelevant-files attributes are deprecated and their behavior
is kept unchanged for backwards-compatibility.

For the motivation and more details see the specification at
https://review.opendev.org/c/zuul/zuul/+/839550

Change-Id: I0a71f4afc82b6a736b1ab13fc7862f23167845f7