Fix merging reconfiguration events
Reconfiguration events for the same tenant are supposed to be merged
so that if many branches are created, only one reconfiguration happens.

However, an error existed in the merge method which raised an exception
preventing the merge, and the exception handler around the merge method
was too broad and therefore suppressed the error.

This is well tested, but with tests that look more like "unit" tests
rather than functional tests.  They did not contain the branch_cache_ltime
dict which triggered the error.

This patch corrects the error and updates the test to be more realistic.

Change-Id: I0b5c7628a0580db55b56eec8c081ebee4d31d989