Fix erroneous semaphore leak cleanup
When a scheduler starts, we run some cleanup routines in case
this is the first scheduler start after an unclean shutdown.  One
of the routines detectes and releases leaked semaphores.  There
were two issues with this which caused us to erroneously detect
a semaphore as leaked and delete it.  Fixing either alone would
have been sufficient to avoid the issue; fixing both is correct.

1) The pipeline state object was not serializing its layout uuid.

This means that every time a new pipeline state object is created,
it is assumed that its layout is newer than what is in zookeeper,
and a re-enqueue must be performed.  All existing queue items are
moved to the old_queues attribute, and a scheduler (possibly a
different one) will re-enqueue them.

2) The semaphore leak detection method did not ask for "old" queue
items.  This means that if a re-enqueue is in progress, it
will only see the queue items which have been re-enqueued.

These two issues combined cause us to fairly reliably move all of
the queue items out of the active queue and then look at only the
active queue to determine which semaphores are in use (which is to
say, none) and then release those semaphores.

This patch corrects both issues.

Change-Id: Ibf5fca03bb3bd33fefdf1982b8245be0e09df567