Fix login button check
We should be storing the capabilities.auth object as auth.info
in redux rather than a second copy of the whole info object.  This
is used in only one place, to check whether the login button should
be displayed.  The error was causing it to never be displayed.

This patch corrects that (and has been tested with multi-tenant,
whitelabel, and sub-path configurations).

Change-Id: I558ecf84f101150465eb5b62bc5787bf9a353793