Display overall duration in buidset page in zuul web
The overall duration is from a user (developer) point of view, how much
time it takes from the trigger of the build (e.g. a push, a comment,
etc.), till the last build is finished.

It takes into account also the time spent in waiting in queue, launching
nodes, preparing the nodes, etc.

Technically it measures between the event timestamp and the end time of
the last build in the build set.

This duration reflects the user experience of how much time the user needs
to wait.

Change-Id: I253d023146c696d0372197e599e0df3c217ef344