Compress ZKObject data
To save space in ZooKeeper compress the data of ZKObjects. This way we
can reduce the amount of data stored in some cases by a factor of 15x
(e.g. for some job znodes).

In case for data that is not yet compressed the ZKObject will fall back
to loading the stored JSON data directly.

Change-Id: Ibb59d3dfc1db0537ff6d28705832f0717d45b632
Depends-On: https://review.opendev.org/c/zuul/zuul/+/789654