Update patch set 2
Patch Set 2:

I thought that section only applied to the "download-commands" plugin (which is a core plugin), and it only controls what is shown to the user and doesn't actually disable any methods.

I'd appreciate it if you would double check that.  I'm working on an issue with someone who has the login-redirect plugin installed.  I suspect the solution to that issue may be to use "a/" in all HTTP git remotes for gerrit in zuul since that works in all cases.  If your site is actually in the same situation, then using "a/" may be a single solution to both problems.  And due to the way gerrit allocates connections, it may be more efficient.

Maybe we should still merge this change, but in general, I'd prefer to keep the number of permutations of how we connect to different systems as small as possible, and gerrit already has far more than any other driver.

If you could try cloning a repo with authenticated http and an "a/" prefix (like: git clone http://gerrit/a/project) that would be great.

Patch-set: 2
Attention: {"person_ident":"Gerrit User 31760 \u003c31760@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}