Update patch set 2
Patch Set 2:

This is the too many open files issue again which we only see in tests running on 3.8; let's see if 3.9 works (that's what I run locally).

Unlike the last time this showed up; I don't think there's anything we can tweak in this change to avoid the issue.  Using zlib (not importing -- actually running a "compress" call) pulls in a lot more shared libraries.  So tests which currently don't actually open them probably are doing so now, and that's correct.

Patch-set: 2