Don't join the command thread if we stop via command socket
If we stop the scheduler via the command socket the stop method fails
with [1] because it tries to join the command thread which is the
thread that runs the stop function. Check for that to make stopping
via 'zuul-scheduler stop' work.

[1] Traceback
ERROR zuul.Scheduler: Exception while processing command
Traceback (most recent call last):
  File "/opt/zuul/lib/python3.8/site-packages/zuul/scheduler.py", line 339, in runCommand
    self.command_map[command]()
  File "/opt/zuul/lib/python3.8/site-packages/zuul/scheduler.py", line 326, in stop
    self.command_thread.join()
  File "/usr/local/lib/python3.8/threading.py", line 1008, in join
    raise RuntimeError("cannot join current thread")
RuntimeError: cannot join current thread

Change-Id: Icf711d5420eebcd24e0ca2c03ad95b416dad1274