Add "zuul delete-pipeline-state" command
This is intended to aid Zuul developers who are diagnosing a bug
with a running Zuul and who have determined that Zuul may be able to
correct the situation and resume if a pipeline is completely reset.

It is intrusive and not at all guaranteed to work.  It may make things
worse.  It's basically just a convenience method to avoid firing up
the REPL and issuing Python commands directly.  I can't enumerate the
requirements where it may or may not work.  Therefore the documentation
recommends against its use and there is no release note included.

Nevertheless, we may find it useful to have such a command during
a crisis in the future.

Change-Id: Ib637c31ff3ebbb2733a4ad9b903075e7b3dc349c