Don't load some drivers in zuul-web, and add keystore reno
Two issues were observed:

* Zuul-web required MQTT connection secrets
* Zuul-web required the keystore password

The first is an oversight in code; when we added the extra
(non-web-based, i.e., Gerrit) source drivers to zuul-web, we
incidentally included drivers like MQTT which still aren't strictly
required.  To correct this, we update the connection initialization
to be a little more specific.

The second is an oversight in documentation.  Zuul-web does use
the keystore to answer requests for public keys now (and we generate
public keys from private keys), so it does legitimately need access
to the keystore.  This change adds a release note to indicate that
(our original release note for the keystore indicated that only the
scheduler and executor require it), and updates the documentation
for zuul-web to indicate it is required.

Change-Id: I4673c28272576e1e5d6d8123a93fb46abfc85348