Cleanup timer triggers in test fixtures
A number of timer triggers in our test fixtures set the time in a list.
This appears to cause Zuul to complain with messages like:

 zuul.ConfigLoader WARNING  1 errors detected during tenant-one tenant configuration loading
 zuul.ConfigLoader WARNING  Zuul encountered a syntax error while parsing its configuration in the
 repo common-config on branch master.  The error was:

   extra keys not allowed @ data['trigger']['timer']

 The error appears in the following pipeline stanza:

   pipeline:
       name: periodic
       manager: independent
       trigger:
         timer:
           - time: '* * * * * */1'
   in "common-config/zuul.yaml@master", line 14, column 3

Address this by removing the extra - to make this a list.

Note this misconfiguration seems to be common in the real world too as
at least OpenDev does it too. We might want to do some communication
about this?

My main motivation for making this change is to remove those attractive
nuisance error warnings from our CI job logs.

Change-Id: Ia46a7ff0bfa72805d87ac6bdccba2fe57ac2ef0f