Update tenant layouts in background
A tenant reconfiguration is when a scheduler encounters an event which
it determines requires creating a new layout.  A tenant layout update is
when a scheduler determines that its local layout is not up to date with
the most recent tenant reconfiguration.  A tenant reconfiguration on one
scheduler is shortly followed by a tenant layout on all other schedulers.

Currently, both of these actions happen in the main scheduler run loop
which means that in the case of a large tenant, one scheduler will dedicate
itself to reconfiguring that tenant while all other schedulers handle all
other tenants (so far, so good).  Then the first scheduler will resume
handling all tenants while all other schedulers simultaneously dedicate
themselves to updating the large tenant.  During this time, only a single
scheduler (the first) will actually be processing pipelines.

We may be able to see more performance out of the schedulers if we allow
them to continue processing other tenants while they update their local
layouts.

This change moves the layout updates into a separate thread which will
run in parallel with the main run loop.

Change-Id: I47dc07338ab814aab2392fc05b1f7656a2a2cbfd