Serialize FrozenJob.affected_projects to ZooKeeper
Currently, this attribute is not serialized to ZooKeeper. Most probably
this is the case because it's only used to schedule the global repo
state (which is usually done within the same pipeline iteration and thus
on the same scheduler).
However, if an exception occurs and the pipeline processing fails, another
scheduler will pick up and try to schedule the global repo state.
Doing so will fail because the affected_projects attribute is not
available on this scheduler as it relies on the data stored in
ZooKeeper:

2022-01-15 07:18:28,621 ERROR zuul.Scheduler: Exception in pipeline processing:
Traceback (most recent call last):
  File "/opt/zuul/lib/python3.8/site-packages/zuul/scheduler.py", line 1746, in _process_pipeline
    while not self._stopped and pipeline.manager.processQueue():
  File "/opt/zuul/lib/python3.8/site-packages/zuul/manager/__init__.py", line 1501, in processQueue
    item_changed, nnfi = self._processOneItem(
  File "/opt/zuul/lib/python3.8/site-packages/zuul/manager/__init__.py", line 1382, in _processOneItem
    ready = self.prepareItem(item)
  File "/opt/zuul/lib/python3.8/site-packages/zuul/manager/__init__.py", line 1322, in prepareItem
    self.scheduleGlobalRepoState(item)
  File "/opt/zuul/lib/python3.8/site-packages/zuul/manager/__init__.py", line 1179, in scheduleGlobalRepoState
    project_cnames.update(job.affected_projects)
AttributeError: 'FrozenJob' object has no attribute 'affected_projects'

For consistency reasons we also serialize the config_hash attribute
which is also missing.

Change-Id: I9425ccf9c532e59b96040d182e2ffa3b5c3321be