Update build set states after scheduling actions
Some of the buildset's state attributes for the files, merge and repo
state where updated before the corresponding action was triggered. This
can lead to hanging build sets in case there is an exception after
updating the state but before or during scheduling of the action.

To solve this we will only update the state attribute(s) after the
action was scheduled. This way the action can be retried in case of an
error.

Change-Id: Ia062d66866d773be4d7c26b6753feb29bf217d0c