Use new layout UUID when resetting pipeline state
In case the layout is updated and the pipeline state needs to be reset,
we need to include the new layout UUID.

Otherwise all schedulers will try to reset the pipeline state when
updating the local layouts. This can lead to broken pipelines in certain
edge szenarios, e.g. resetting the state while the pipeline is currently
processed.

Change-Id: I01ad30b635d8930587ba1e4fbc2844a384753303