autohold: use canonical project name in REST interface
I noticed autoholds not matching as I expected when using the
zuul-client like

 $ zuul-client autohold --tenant openstack \
   --project opendev/system-config --job ...

It turns out the autohold matching does a full match against the
canonical project name when deciding if a hold matches.

The gearman RPC client expands your --project argument to it's
canonical name @ zuul/rpclistener.py:RPCListener.handle_autohold()

 (trusted, project) = tenant.getProject(args['project'])
  if project:
      params['project_name'] = project.canonical_name
  else:
      error = "Invalid project: %s" % args['project']

This does the same in the REST client side, so similarly formatted
autohold commands from zuul-client work the same as for the gearman
client.

Change-Id: Id565d540cbd6a5a2108b49af5db2c2e308b26cc3