Revert "Fix a bug in getting changed files"
The reverted change can lead to the listing of files that are not
changed in the referenced commit(s). This can e.g. happen if the base
branch (e.g. master) has diverged from the feature branch.

This is now also tested to avoid regressions in the future. The issue
related to files that are added/removed in the same range of commits
(e.g. a PR) needs to be addressed in a separate change.

This reverts commit e63d7b0cdb4a38fd2abbd962e77800dac8c2bd22.

Change-Id: I07bc4a09bf162fdbc4c2daeecb19e12d81241801