Fix update of mutable ZKObject attributes
In cases where we modify mutable ZKObject attributes and try to write
them out via ZKObject.updateAttributes() it can happen that the state is
not written out to Zookeeper.

The reason for that is the check of the old vs. new serialized data of
the ZKObject in order to determine if the object was indeed updated.

Since the instance of the updated attribute is already part of the
ZKObject instance the version used in the old and the new state is the
same.

As far as I could see this was only a problem in the case of a bundle
(used for circular dependencies) and for the relative priority queues.

Change-Id: I81afcf1b65c66fda54e24cf4b8c04b5b2608f8f1