Use hash to de-duplicate ZKObject writes
We currently serialize ZKObjects twice, once before and once after
perofrming our local attribute updates in order to determine whether
we should write the object to ZK.  This works except in the case of
a mutable attribute which was mutated outside of the update method.
That happens in two places: building change queues and item bundles.
In these cases, the two serialized values will be the same.

To correct this, store a hash of the data we read from ZooKeeper, and
on writing to ZK, compare the hash of the new serialized value to that.
Only write to ZK if it differs.

Additionally, because we share a bundle object between queue items,
make sure that we go back and write all items in a bundle to ZK if we
add a new item.

Change-Id: I234833a707aeb41aea94e7bd653c1a94365ae4dc
Co-Authored-By: Simon Westphahl <simon.westphahl@bmw.de>