Prevent run handler and layout update deadlock
Since the run handler and the layout update thread lock the layout and
tenant in a different order, we can run into a deadlock that will block
the run handler on a scheduler. The affected tenant can also no longer
be processed by any other scheduler.

To fix this, we will no longer block on the layout lock. This is analog
to how we are already using the tenant read lock. The alternative would
be to switch the locks around so we are always acquiring them in the
same order.

The observed deadlock looked as follows:
* run handler: doTenantReconfigureEvent(): acquire layout lock
* layout thread: runHandlerLayoutUpdates(): acquire tenant read lock
* run handler: doTenantReconfigureEvent(): block on tenant write lock
*layout thread: updateTenantLayout(): block on layout lock

Change-Id: Ic4c169692ffcbc05da4586a5b1f8be5ce5397f0f