Update patch set 1
Patch Set 1: Code-Review-1

Can you confirm that you are just running "./test-setup-docker.sh" from the tools directory?

I do that quite often, and it works fine with the ZK 3.7.0 docker images.

Docker mounts all of the volumes before starting the process, and the certs are already created, so I don't see how that can be causing the problem.

The reason this is needed in the tutorial is that docker-compose starts all the containers first, and one of the containers is responsible for creating the certificates.  Therefore the ZK container starts without any certs available.  That's not the case with test-setup-docker.

Patch-set: 1
Label: Code-Review=-1