Update patch set 1
Patch Set 1: Code-Review-1

As mentioned in https://review.opendev.org/817383 I'd like to avoid this for now; it's not clear that this will be a win, so I'd like to get a better baseline first and then think about how we want to evaluate it.

This could improve performance in some circumstances, but it might actually make it worse when it's very busy (because every time we skip a locked tenant, it's a win).

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1