Cache serialized tenant status
We see a very high cpu load on zuul web. Profiling shows that most of
that time is spent during json serialization. The biggest json blobs
we have is the per tenant status json. This is currently cached as
data structure so we can filter it by change as well. In order to
reduce json serialization effort also cache a pre-serialized version
of this and use it for the tenant status.

Change-Id: Idbbcda6bf0835fadf970e4fb43adc300770da8e7