Fix bootstrapping zk data model version
On startup, we would not wait for all components to be registered
before we decided what the minimum model version is.  To correct
that, wait until the initial bootstrap sequence is complete before
setting the version (this is the self._init=True part of the change).

Additionally, we initialized the component data structure such that
the model_api was *our* model api.  This meant that registration
automatically set our correct api version.  However, it meant that
we would read any other component in zk without a model_api entry
and initialize it to our value.  This means that the 0->1 model
transition never happens and we assume every component has at least
version 1.

That's harmless in practice since 0==1, but it's still incorrect
and forms a bad pattern for future attribute additions.  Correct this
by initializing it to 0 and then updating it to our version on
registration.

Change-Id: I830301c652ea603fa8c6f1f800a3a29e9f3ce060