Fix version check when re-creating cached changes
When a change is removed from the cache and added back without all
schedulers noticing the removal (e.g. due to a GC pause), we can end up
with old versions of a change in the cache that might not be updated
correctly.

The reason for this is that we are using the ZNode version to decide if
a change is already up-to-date. As the ZNode version is reset to 0 when
the node is recreated we failed to detect such updates.

This can lead to the cleanup of data nodes that are still in use but are
not referenced from the local cache due to the issue described above.
With that the cached change will be removed eventually (due to a
'corrupt' data node) and we might end up with a broken pipeline state as
the change key can no longer be resolved.

To fix this we can use the last modified transaction ID of the ZNode to
detect if a change was updated, as the mzxid is monotonically
increasing.

Change-Id: I85f388cfbad04cf82deaf5cc8313d389eae354e4