Explicitly close finger sockets in web
If a user opens a websocket console stream and then closes it, we
unregister the log streamer from the stream handler which means
that it will no longer be serviced in the poll loop.  The object
will also be garbage collected.  But until that happens, the
finger socket remains open and using resources.  Explicitly close
it in that condition to speed cleanup.

Change-Id: I6da1c88dfe20eb9e1ada09d7aa741f9024ddfc04