Identify cherrypy requests in logs
This adds a unique ID to every cherrypy request and adds it to the
all of the log lines in the web server.

Some requests produce multiple log entries, and this will allow us
to associate them with each other as well as the cherrpy access log
which is emitted at the end of the request.

To add this to cherrypy's request logging, we subclass the cherrypy
LogManager class and add accessor methods so that whenever the
superclass requests a logger, it gets our annotated logger.

See: https://github.com/cherrypy/cherrypy/blob/main/cherrypy/_cplogging.py

There is an underscore in that module name, but several parts of the
cherrpy documentation suggest importing underscore modules for the
purpose of extending cherrpy.

On the ZuulApi class, we add a similar accessor method for obtaining
the annotated logger for any request.

The two log streaming classes initialize their log attribute in the
constructor since their instances are dedicated to a single request.

The remaining classes retain the class-level logger since they operate
outside of the request context.

Change-Id: I66af717ab34fa4ad3492a01badb25e3865133104