Don't delete items of partially dequeued bundles
In case the queue processor is interrupted after dequeuing an item that
is part of a bundle, we can end up in a situation where an item is still
needed but no longer part of any change queue.

Due to that we need to also consider items that are still part of a
bundle when deciding which items can be cleaned up.

This issue can lead to a blocked pipeline as the state can no longer be
refreshed from ZK after the cleanup.

Change-Id: Ib69a7733f63f50eb2105de87653196e6da25c887