Don't lock pipelines during local layout update
In our Zuul deployment we can see that the local layout update takes
very long on a busy scheduler. We also observed that due to the locking
in the post-config hook, we might starve pipelines for the duration of
the local layout updates.

Since the pipeline state is already reset after the inital
reconfiguration we don't need to reset it again after a local layout
update. In case the pipeline's layout UUID is already the same as our
current one, we can skip that step.

Because we no longer restore the pipeline state from ZK in _postConfig,
we may exit the _reconfigureTenant method with an out of date copy of
the queue contents.  In particular, on startup our in-memory queue may
be empty.  Because of that, remove the statsd emission in that method
and rely more heavily on the pipeline manager emitting those stats.
This statsd emission was most important on scheduler startup with an
empty state, which is not as relevant any more now that state persists
in ZK.

This change also removes the relative priority queues from the pipeline
state. The relative prio queues were never properly (de-)serialized and
shouldn't be stored in ZK at all.

This also corrects a use of locks in the prime method.  This case is
probably impossible to hit with the current code.  The lock which is
passed to the ZKContext in prime is only used if _reconfigureTenant
needs to modify objects in ZK for newly created tenants, but a newly
created tenant has no ZK objects that need modifying.  Because of that,
a test for this code is unfeasible, but let's correct the error anyway.
The lock_ctx is a generator, and the result of entering it as a context
manager is the actual lock that we need to pass to the ZKContext.

Change-Id: I7ad003a8fc24ab9be417a46287390675227e2db3